+++ objectbox/CHANGELOG.md
@@ -4,6 +4,7 @@
   `objectbox_flutter_libs` (and `objectbox_sync_flutter_libs`). Use this to fix loading the 
   ObjectBox library on these devices. [#369](https://github.com/objectbox/objectbox-dart/issues/369)
 * Improve code generator performance if there are many entities with many constructor parameters.
+* Throw `StateError` instead of crashing on closed `Box`, `Query` and `PropertyQuery`.
 
 ## 2.1.0 (2023-06-13)